Add document about open source strategy
This is a refined version of [issue 160](https://github.com/dtr-org/unit-e/issues/160).

The next required actions have been extracted to GitHub issues so
we can prioritize and plan them there on a project board. There are
links to the issues in the documents in the corresponding sections.

This document is supposed to provide the context and the overall
summary of what we are doing there. Some of the content will
probably move to the actual documentation of what we are doing in
README.md or CONTRIBUTING.md, for example.